+++ compiler-core/src/language_server.rs
@@ -3,6 +3,7 @@ mod compiler;
 mod engine;
 mod feedback;
 mod files;
+mod messages;
 mod progress;
 mod router;
 mod server;
@@ -16,7 +17,8 @@ use crate::{
     ast::SrcSpan, build::Target, line_numbers::LineNumbers, manifest::Manifest,
     paths::ProjectPaths, Result,
 };
-use lsp_types::{Position, Range};
+use camino::Utf8PathBuf;
+use lsp_types::{Position, Range, Url};
 use std::any::Any;
 
 #[derive(Debug)]
@@ -42,4 +44,14 @@ pub fn src_span_to_lsp_range(location: SrcSpan, line_numbers: &LineNumbers) -> R
         Position::new(start.line - 1, start.column - 1),
         Position::new(end.line - 1, end.column - 1),
     )
+}
+
+fn path(uri: &Url) -> Utf8PathBuf {
+    // The to_file_path method is available on these platforms
+    #[cfg(any(unix, windows, target_os = "redox", target_os = "wasi"))]
+    return Utf8PathBuf::from_path_buf(uri.to_file_path().expect("URL file"))
+        .expect("Non Utf8 Path");
+
+    #[cfg(not(any(unix, windows, target_os = "redox", target_os = "wasi")))]
+    return Utf8PathBuf::from_path_buf(uri.path().into()).expect("Non Utf8 Path");
 }
